Refactor : socket 패키지에 dto 패키지 생성
socket 패키지에 파일이 너무 많이 내부 패키지에 속해있지 않고 나와있어 dto 패키지를 생성합니다.
